Finish Delete

+++ src/components/table/ListTicketByMatch.jsx
@@ -4,9 +4,10 @@ import { ticketColumns } from "../../datatablesource";
 import { useState, useEffect } from "react";
 import axios from "../../AxiosConfig";
 import LoadingSpinner from "../../pages/LoadingWait/LoadingSpinner";
+import swal from 'sweetalert';
+import Swal from "sweetalert2";
 
 var path = "match/tickets/"
-
 const ListStadium = () => {
     var matchId = localStorage.getItem("idClickTicketByMatch")
     const [isRender, setisRender] = useState(true);
@@ -34,18 +35,39 @@ const ListStadium = () => {
         return window.location.href = "../ticket/updateTicket"
     };
 
-    //Handle Sold Out here ----------------------------------------------------------------------------
+    //Handle Delete here ----------------------------------------------------------------------------
+    function showError(text) {
+        Swal.fire({
+            title: "Oops...",
+            text: text,
+            icon: "error",
+            confirmButtonText: "OK",
+        })
+    }
+
     const handleDelete = (id) => {
-        console.log(id);
-        axios.delete(path + matchId + "/" + id)
-            .then(res => {
-                console.log("check delete ", res);
-                alert('Deleted ticket by id: ' + id);
-                setData(data.filter((item) => item.id !== id));
-                window.location.reload();
-            })
-            .catch(function (err) {
-                console.log(err);
+        swal({
+            title: "Are you sure?",
+            text: "Once deleted, you will not be able to recover this " + id,
+            icon: "warning",
+            buttons: true,
+            dangerMode: true,
+        })
+            .then((willDelete) => {
+                if (willDelete) {
+                    axios.delete(path + matchId + "/" + id)
+                        .then(res => {
+                            swal("Poof! " + id + " has been deleted!", {
+                                icon: "success",
+                            });
+                            window.location.reload();
+                        })
+                        .catch(function (err) {
+                            showError(err);
+                        });
+                } else {
+                    swal(id + " is safe!");
+                }
             });
     };
